Fix #4, #5: Optional 1:1 and 1:many edges (#40)
This is one of the highly requested features, but even more importantly it paves a way for fixing #18 .

![Screenshot 2024-10-28 at 21 43 23](https://github.com/user-attachments/assets/41f323e2-061c-4e7f-809c-27b78f96cd07)
![Screenshot 2024-10-28 at 21 44 24](https://github.com/user-attachments/assets/36bf76ee-3b84-4e57-be31-8a090ca36981)

No impact on cascading deletes yet, the logic stays the same (deletion propagates in the direction towards the foreign key, but not from it). This will not work for #18, so additional configuration will be needed to reverse/allow both directions.

----

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.